# [libc] memmove optimizations