@dvyukov 2023-10-24
Oh, "Rebase and merge" is not enabled, that's sad.
Do you know if I am supposed to apply this to main branch locally and try to push until it succeeds? Or there is a better way?