@dvyukov 2023-10-25
> A slightly cleaner way to do it may be to return true/false from inline_memmove

Does it look better?

> Is this the logic you meant to implement? 

Yes.

Does the following look better?
```
  if (vector_size >= 32 ? count < 32 : count <= 32)
    return generic::Memmove<uint128_t>::head_tail(dst, src, count);
```