@gchatelet 2023-10-25
I missed your reply before sending mine...

> Does the following look better?
> 
> ```
>   if (vector_size >= 32 ? count < 32 : count <= 32)
>     return generic::Memmove<uint128_t>::head_tail(dst, src, count);
> ```

It's more explicit yes.