@dvyukov 2023-10-26
Moved the benchmarks change to #70302.
And I am going to squash the 2 remaining memmove changes into 1 for now.